+++ packages/types/src/nil.ts
@@ -1,14 +1,20 @@
 /**
  * A value that can be of type T or undefined.
+ *
+ * @since 0.2.1
  */
 export type Maybe<T> = T | undefined;
 
 /**
  * A value that can be of type T or null.
+ *
+ * @since 0.2.1
  */
 export type Nullable<T> = T | null;
 
 /**
  * A value that can be of type T, null, or undefined.
+ *
+ * @since 0.2.1
  */
 export type Nilable<T> = T | null | undefined;